cli: refactor to several files
