+++ package.json
@@ -10,8 +10,8 @@
     "build": "tsc",
     "lint": "tslint --project \"./tsconfig.json\"",
     "test": "jest",
-    "testdata:generate": "ts-node --files ./src/cli.ts generate",
-    "testdata:validate": "ts-node --files ./src/cli.ts validate",
+    "testdata:generate": "ts-node --files ./src/cli/cli.ts generate",
+    "testdata:validate": "ts-node --files ./src/cli/cli.ts validate",
     "protobuf": "node_modules/.bin/pbjs -t static-module -w commonjs -o src/proto/chainscript_pb.js proto/chainscript.proto && node_modules/.bin/pbts -o src/proto/chainscript_pb.d.ts src/proto/chainscript_pb.js"
   },
   "files": [
