+++ src/cli/test_case.ts
@@ -1,0 +1,11 @@
+/**
+ * Every test case should implement the ITestCase interface.
+ */
+export interface ITestCase {
+  /** Id of the test case. */
+  id: () => string;
+  /** Generate encoded segment bytes. */
+  generate: () => string;
+  /** Validate encoded segment bytes. */
+  validate: (encodedSegment: string) => void;
+}
